refactor: isolate SearchField stories
